+++ app/src/main/res/layout/activity_container.xml
@@ -3,16 +3,15 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
-    <Button
-        android:id="@+id/btn_change"
-        android:layout_width="match_parent"
-        android:layout_height="50dp"
-        android:text="更换Fragment " />
+<!--    <Button-->
+<!--        android:id="@+id/btn_change"-->
+<!--        android:layout_width="match_parent"-->
+<!--        android:layout_height="50dp"-->
+<!--        android:text="更换Fragment " />-->
 
     <FrameLayout
         android:id="@+id/fl_container"
         android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:layout_below="@+id/btn_change"/>
+        android:layout_height="match_parent"/>
 
 </RelativeLayout>
